refactored: playwright's tests
