+++ test/conftest.py
@@ -1,4 +1,5 @@
 import pytest
+from shared.engine import get_cookies, get_proxy_setting
 from shared.mock_html import get_corrupted_html, get_html
 from shared.mock_json import get_json_data
 
@@ -16,3 +17,13 @@ def corrupted_html():
 @pytest.fixture
 def json_data():
     return get_json_data()
+
+
+@pytest.fixture
+def proxy_setting():
+    return get_proxy_setting()
+
+
+@pytest.fixture
+def cookies():
+    return get_cookies()
